# perf: compute subtrees in parallel when computing proof