--- a/ssz-rs/Cargo.toml
+++ b/ssz-rs/Cargo.toml
@@ -32,6 +32,11 @@ snap = "1.0"
 project-root = "0.2.2"
 serde_json = "1.0.81"
 hex = "0.4.3"
+criterion = { version = "0.5", features = ["html_reports"] }
 
 [build-dependencies]
 sha2 = "0.9.8"
+
+[[bench]]
+name = "compute_proof"
+harness = false