--- a/ssz-rs/Cargo.toml
+++ b/ssz-rs/Cargo.toml
@@ -13,12 +13,14 @@ exclude = ["tests/data"]
 
 [features]
 default = ["serde", "std"]
+hashtree = ["dep:hashtree"]
 std = ["bitvec/default", "sha2/default", "alloy-primitives/default"]
 sha2-asm = ["sha2/asm"]
 serde = ["dep:serde", "alloy-primitives/serde"]
 
 [dependencies]
 bitvec = { version = "1.0.0", default-features = false, features = ["alloc"] }
+hashtree = { version = "0.2.0", optional = true, package = "hashtree-rs" }
 ssz_rs_derive = { path = "../ssz-rs-derive", version = "0.9.0" }
 sha2 = { version = "0.9.8", default-features = false }
 serde = { version = "1.0", default-features = false, features = [