Merkle trees are binary trees, enabling parallel processing of subtrees. By parallelizing 8 subtrees, we get ~83% reduction in processing time for two mainnet blocks. This implementation detects the number of available CPU cores and runs either serial processing, parallel 4 subtrees (RPi4) or 8 subtrees.

Process 2 subtrees in parallel
```
Prove Benchmark - File: benches/21315748.json - size 247/123
                        time:   [250.73 ms 251.78 ms 253.11 ms]
                        change: [-44.361% -43.919% -43.531%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) high mild
  1 (10.00%) high severe

Prove Benchmark - File: benches/21327802.json - size 261/130
                        time:   [251.22 ms 252.05 ms 252.89 ms]
                        change: [-44.012% -43.797% -43.578%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Process 4 subtrees in parallel compared to 2
```
Prove Benchmark - File: benches/21315748.json - size 247/123
                        time:   [158.15 ms 159.54 ms 161.30 ms]
                        change: [-42.477% -41.625% -40.843%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking Prove Benchmark - File: benches/21327802.json - size 261/130: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 8.7s or enable flat sampling.
Benchmarking Prove Benchmark - File: benches/21327802.json - size 261/130: Collecting 10 samples in estima
Prove Benchmark - File: benches/21327802.json - size 261/130
                        time:   [158.71 ms 159.15 ms 159.40 ms]
                        change: [-43.690% -43.042% -42.435%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 10 measurements (30.00%)
  2 (20.00%) low mild
  1 (10.00%) high severe
```

Process 8 subtrees in parallel compared to 4
```
Prove Benchmark - File: benches/21315748.json - size 247/123
                        time:   [92.659 ms 93.286 ms 93.841 ms]
                        change: [-41.520% -40.689% -39.987%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking Prove Benchmark - File: benches/21327802.json - size 261/130: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.2s or enable flat sampling.
Prove Benchmark - File: benches/21327802.json - size 261/130
                        time:   [93.627 ms 94.510 ms 95.063 ms]
                        change: [-40.912% -40.329% -39.725%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Avoid creating extra buffers for subtrees
```
Benchmarking Prove Benchmark - File: benches/21315748.json - size 247/123: Collecting 10 samples in estima
Prove Benchmark - File: benches/21315748.json - size 247/123
                        time:   [79.151 ms 79.399 ms 79.855 ms]
                        change: [-14.591% -13.868% -13.113%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking Prove Benchmark - File: benches/21327802.json - size 261/130: Collecting 10 samples in estima
Prove Benchmark - File: benches/21327802.json - size 261/130
                        time:   [82.095 ms 83.493 ms 84.412 ms]
                        change: [-14.100% -12.710% -11.136%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Hasher abstractions from https://github.com/ralexstokes/ssz-rs/pull/161