@estensen 2025-08-06
superseeded by https://github.com/ralexstokes/ssz-rs/pull/177
hashing is not the bottleneck